+++ ConsentViewController/Classes/Consents/SPGDPRConsent.swift
@@ -67,6 +67,19 @@ public typealias SPGDPRPurposeId = String
     /// that's the internal Sourcepoint id we give to this consent profile
     public var uuid: String?
 
+    /// a list of ids of the categories accepted by the user in all its vendors.
+    /// If a category has been rejected in a single vendor, its id won't part of the `acceptedCategories` list.
+    public var acceptedCategories: [String] {
+        let categoryGrants = vendorGrants
+            .flatMap { $0.value.purposeGrants }
+        return Set<String>(categoryGrants.map { $0.key })
+            .filter { id in
+                categoryGrants
+                    .filter { $0.key == id }
+                    .allSatisfy { $0.value }
+            }
+    }
+
     public init(
         uuid: String? = nil,
         vendorGrants: SPGDPRVendorGrants,